[FEAT] PlayerIndexingTest, PlayerServiceTest 분리 작업
