+++ src/test/java/com/wearei/finalsamplecode/domain/player/service/PlayerServiceTest.java
@@ -1,0 +1,48 @@
+package com.wearei.finalsamplecode.domain.player.service;
+
+import com.wearei.finalsamplecode.domain.player.dto.response.PlayerSearchResponse;
+import com.wearei.finalsamplecode.domain.player.entity.Player;
+import com.wearei.finalsamplecode.domain.player.repository.PlayerRepository;
+import com.wearei.finalsamplecode.domain.team.entity.Team;
+import com.wearei.finalsamplecode.domain.team.repository.TeamRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+import java.util.List;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+
+@Transactional
+@SpringBootTest
+public class PlayerServiceTest {
+    @Autowired
+    private TeamRepository teamRepository;
+
+    @Autowired
+    private PlayerService playerService;
+
+    @Autowired
+    private PlayerRepository playerRepository;
+
+    @BeforeEach
+    public void setUp() {
+        Team team = teamRepository.save(new Team("두산베어스", "url1", "url2", "url3", "url4"));
+        Player player = playerRepository.save(new Player(1L, 20L, "박서연", "두산베어스", "외야수", "https://youtube.com", "Normal", 123L));
+    }
+
+    @Test
+    public void testGetPlayerByName() {
+        // Given
+        String playerName = "박서연";
+
+        // When
+        List<PlayerSearchResponse> players = playerService.getPlayerByNameAndTeamName(playerName, null);
+
+        // Then
+        assertFalse(players.isEmpty(), "선수 정보가 조회되어야 합니다.");
+        assertEquals(playerName, players.get(0).getPlayerName(), "조회된 선수 이름이 일치해야 합니다.");
+        assertEquals("두산베어스", players.get(0).getTeamName(), "조회된 팀 이름이 일치해야 합니다.");
+    }
+}
